requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
channels==3.0.4
paho-mqtt == 2.0.0